export the sendmessage to pull owner to another function
